requirements.txt
discord.py
aiohttp
PyYAML

